requirements.txt
requests
python-dotenv
pyairtable
apify-client
playwright
beautifulsoup4
lxml
flask
flask-cors
